Now effects are attached into effects file and is imported into main index.js
